socket received a message on given bytes
Before, the socket received always on 1024 bytes, but it could cause problems when multiple messages are send consecutively. Like, the server send the permission to a group, the allow th join the group. Those 2 different action could be received as a single message. So now, the size of the message is send just before the actual message