add the group freature and formate message before sending
The format is a dictionary with the content, the sender and the target (no target in the server formating). On connection, the client receive the list of existing groups, gain access to the default predefined group, then join it. The sended message is now broadcast based on the target. The target is the group where the client is.