add preconnection protocole
Sorry, it was too hard to split that in smaller commits without have commits where the program is broken. So, there is a new EntryForFormatedMessage for the nickname, a new client action to request a connection, and server actions to give a temporary nickname and accept a connection. In server.py, the client receive an id as a new data, each client have a unique one. On this new protocole, the client create the socket at the start of the program, where the server give him a temporary nickname (his unique id from server side). The connexion protocole if modifie in the information send by the client, it now give the nickname gived by the user. Look at the ticket #10 on the original repo for better understanding