reconnection protocole: complete
add an action from the server "acceptReconnection" and a new error type "alreadyConnected". Now, on a connection request, the server, if the submitted name exists, check if it is connected (refuse connection here), else it considere the request as a reconnection. Yeah, anybody can steal the account of a disconnect client. Maybe a password feature incoming...